Development

+++ lib-network/src/apps/tftp/tftpdaemon.cpp
@@ -2,7 +2,7 @@
  * @file tftpdaemon.cpp
  *
  */
-/* Copyright (C) 2019-2023 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2019-2024 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -304,7 +304,7 @@ void TFTPDaemon::HandleRecvData() {
 		m_nDataLength = m_nLength - 4;
 		m_nBlockNumber = __builtin_bswap16(pDataPacket->BlockNumber);
 
-		DEBUG_PRINTF("Incoming from " IPSTR ", m_nLength=%u, m_nBlockNumber=%d, m_nDataLength=%u", IP2STR(m_nFromIp), static_cast<uint32_t>(m_nLength), m_nBlockNumber, static_cast<uint32_t>(m_nDataLength));
+		DEBUG_PRINTF("Incoming from " IPSTR ", m_nLength=%u, m_nBlockNumber=%d, m_nDataLength=%u", IP2STR(m_nFromIp), m_nLength, m_nBlockNumber, m_nDataLength);
 
 		if (m_nDataLength == FileWrite(pDataPacket->Data, m_nDataLength, m_nBlockNumber)) {
